pyproject.toml
[tool.poetry]
name = "clash-parser"
version = "0.1.0"
description = "an implementation of the Parser feature from CFW, but with HTTP API"
authors = ["xyqyear <xyqyear@gmail.com>"]
readme = "README.md"
packages = [{include = "clash_parser"}]

[tool.poetry.dependencies]
python = "^3.11"
ruamel-yaml = "^0.17.31"
aiofiles = "^23.1.0"
aiohttp = "^3.8.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"
coverage = "^7.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
